# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Thesis Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /theses:
    get:
      operationId: listTheses
      tags: [theses]
      summary: List theses
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: studentId
          in: query
          schema:
            type: integer
        - name: supervisorId
          in: query
          schema:
            type: integer
        - name: reviewerId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of theses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Thesis"
    post:
      operationId: createThesis
      tags: [theses]
      summary: Create a thesis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThesisInput"
      responses:
        "201":
          description: Created thesis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thesis"
  /theses/{id}:
    get:
      operationId: getThesis
      tags: [theses]
      summary: Get thesis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Thesis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thesis"
    patch:
      operationId: updateThesis
      tags: [theses]
      summary: Update thesis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThesisUpdate"
      responses:
        "200":
          description: Updated thesis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thesis"
    delete:
      operationId: deleteThesis
      tags: [theses]
      summary: Delete thesis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /theses/{id}/submit:
    post:
      operationId: submitThesis
      tags: [theses]
      summary: Submit thesis for review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Thesis submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thesis"
  /theses/{id}/status:
    patch:
      operationId: updateThesisStatus
      tags: [theses]
      summary: Update thesis status (admin/supervisor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThesisStatusUpdate"
      responses:
        "200":
          description: Updated thesis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thesis"
  /theses/{id}/assign:
    post:
      operationId: assignThesis
      tags: [theses]
      summary: Assign supervisor or reviewer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThesisAssignment"
      responses:
        "200":
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thesis"
...
  /reports/theses:
    get:
      operationId: getThesesReport
      tags: [dashboard]
      summary: Get thesis statistics report
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThesesReport"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [student, supervisor, reviewer, committee_member, admin]
        faculty:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - createdAt
...
    UserInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [student, supervisor, reviewer, committee_member, admin]
        faculty:
          type: string
        department:
          type: string
        phoneNumber:
          type: string
      required:
        - email
        - password
        - firstName
        - lastName
        - role
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        faculty:
          type: string
        department:
          type: string
        phoneNumber:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [student, supervisor, reviewer, committee_member, admin]
    Thesis:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, submitted, under_review, approved, rejected, defended]
        studentId:
          type: integer
        supervisorId:
          type: ["integer", "null"]
        reviewerId:
          type: ["integer", "null"]
        defenseId:
          type: ["integer", "null"]
        keywords:
          type: ["string", "null"]
        field:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        student:
          $ref: "#/components/schemas/User"
        supervisor:
          type: ["object", "null"]
        reviewer:
          type: ["object", "null"]
      required:
        - id
        - title
        - status
        - studentId
        - createdAt
        - updatedAt
        - student
    ThesisInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        keywords:
          type: string
        field:
          type: string
      required:
        - title
    ThesisUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        keywords:
          type: string
        field:
          type: string
    ThesisStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [draft, submitted, under_review, approved, rejected, defended]
        comment:
          type: string
      required:
        - status
    ThesisAssignment:
      type: object
      properties:
        role:
          type: string
          enum: [supervisor, reviewer]
        userId:
          type: integer
      required:
        - role
        - userId
...
    ThesesReport:
      type: object
      properties:
        totalCount:
          type: integer
        byStatus:
          type: object
          additionalProperties:
            type: integer
        byField:
          type: object
          additionalProperties:
            type: integer
        byMonth:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              count:
                type: integer
            required:
              - month
              - count
      required:
        - totalCount
        - byStatus
        - byField
        - byMonth
...