# lib/api-zod/src/generated/types/thesisUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thesis Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ThesisUpdate {
  title?: string;
  description?: string;
  keywords?: string;
  field?: string;
}